Added FizzyUtils to dependency list
So uhh I spent a whole week wondering why I couldn't get this mod to work after I did a fresh reinstall of my primary mods. Turns out I also needed to copy the FizzyUtils dll. The people over at the discord group were nice enough to point out that this is mentioned in the Discord release channel, but it might also need to be here in the github readme in case someone else scratches their head for a week like I did.